nginx: prepare for downgrade
This adds a service which creates a directory needed in case the BMC is
downgraded to a release which used nginx.
The directory must be created in the overlay filesystem (not in the
generated file system) so it will survive the downgrade.  The service
is "harmless" and only creates one additional subdirectory.

This is needed for OpenBMC 2.7 and later systems (which use BMCWeb) and
may be code updated (downgraded) to OpenBMC 2.6 and earlier releases (which
use nginx).

Without this fix, the BMC fails to start its nginx server and does not
provide HTTP, REST, or Web services.  The underlying cause of the server
failing to start is: the directory it needs to write files into does not
is not present.
The problem does not affect the SSH service.  Commands like "systemctl
status nginx" show messages about "unable to write 'random state'".
The problem can be worked around by creating the /etc/ssl/certs/nginx/
directory, then starting the service: systemctl start nginx

This service can be removed in the first release that does not support a
downgrade path to an OpenBMC 2.6 or earlier release.

Tested on an OpenBMC 2.7 image:
  rm -r /etc/ssl/certs/nginx/   # cleanup workarounds
  code update to image which has this fix
  systemctl --no-pager list-units nginx-prep-downgrade.service
  systemctl status nginx-prep-downgrade.service
  ls /etc/ssl/certs/nginx
  The service successfully creates the directory.

Fixes: https://github.com/openbmc/openbmc/issues/3564

Signed-off-by: Joseph Reynolds <jrey@linux.ibm.com>
Change-Id: I32c3231fd877f5f5ea523d584d34892610b6cd0c