meta-ibm: witherspoon-128: MACHINEOVERRIDES fixup
The ordering of the values in MACHINEOVERRIDES is wrong:
   MACHINEOVERRIDES="witherspoon-128:witherspoon"

This means that:
   FOO_witherspoon = "bar"

will override:
   FOO_witherspoon-128 = "baz"

which is the opposite of expectation.  The ordering is a function of
config file include order and choice of the predot vs postdot operator.
Use the same include ordering and dot operators as meta-fsl-arm to get
the overrides to appear in MACHINEOVERRIDES in the correct order such
that the expected semantics are provided and:
  FOO_witherspoon-128 = "bar"

will override:
  FOO_witherspoon = "baz"

(From meta-ibm rev: a12ff8392546591d8b52927da983dd76d2c0f1bd)

Change-Id: I823e236bd0147cc058052d85778b522e74fabb28
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>