wspoon: hwmon: Use labels for finding VRM I/V/P
A recent kernel update added a second input current/voltage/power
reading to the hwmon sysfs attributes for the ir35221 VRMs, and it threw
off the labels in the config files since they were hardcoded for the
specific previous numbers.  For example, curr2_input previously
represented an output current but now represents an input current.

Fix this by using the LABEL mode in hwmon to find which sysfs file to
use for a particular sensor by the contents of its corresponding label
file.  For example, the p0_vdd_current sensor should look at the sysfs
file that has a corresponding 'iout1' value in the currX_label file.

Tested:  Test that the voltages are back within thresholds again and not
showing values around 12V.

Change-Id: Ia01f5213a2e2e84a8e116b146647ffadfc02c0ab