Add service to expire password on first boot
Ensure the root password is set to expired on the first boot of the
system. This can be after the BMC flash chip is first physically written
or after a factory reset.

The dropbearkey.service dependency is to ensure this service does not
run if a user is updating from a OpenBMC firmware image that did not
have this feature enabled to one that does. The user password should not
be reset in this case so use the dropbear rsa key file to know if the
system was running previous firmware (i.e. not a factory reset
scenario).

By default this feature is not enabled.

See the following design for more information:
https://github.com/openbmc/docs/blob/master/designs/expired-password.md

Tested:
- Enabled feature and ensured in QEMU I was required to change the
  password on first boot
- Rebooted QEMU and verified I was not required to change the password
  and the one I had set previously worked

Change-Id: I5abb190df55ee1238f51916442581461766e3cc8
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>