avahi: Ensure user and group added
This is a patch on top of a change already in OP940 that was incorrect
due to a misplaced line feed. This commit updates the code to what is
now in gerrit at:
https://gerrit.openbmc-project.xyz/c/openbmc/meta-phosphor/+/24271

The extra line feed was added to the bbappend due to git requiring some
sort of change to detect the file permission change. git only detects
non-exe to/from exe. This permission change was required to avoid a
warning by systemd.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>