openbmc: Add systemd service for croserver
The upstream application does not ship a systemd service file so we
include it in meta-openpower.

The service does not have an 'install' section as it should be started
on demand by the cronus user. This may change in the future if OpenBMC
gains the ability to put itself in a debug state, where starting the
croserver will be part of that process.

(From meta-openpower rev: c6c86e4602e7d8734e771ee4bf2fb1049acbfb75)

Change-Id: I088ae4acfb7a8f039fd9342ba2314255ec7987f1
Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>