dropbear ssh server expired password dialog
This adds a patch from the dropbear email list.  That patch was modified
to work with DROPBEAR_2019.78.  The patch enhances the dropbear SSH
server to respond to SSH client requests to change the password.

The /etc/pam.d/dropbear config file is enhanced so PAM will allow
dropbear to change the password.  The /etc/pam.d/common-auth file is
enhanced to tolerate the initial SSH auth failure and allow the password
change to succeed.

tested: yes, via the OpenSSH client.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>