meta-ibm: add simicsfs packagegroup
Add a simicsfs packagegroup to packagegroup-base when both the simicsfs
MACHINE_FEATURE and DISTRO_FEATURE are enabled.  This is how oe-core
handles other filesystems.  To enable simicsfs the followiing must
occur:

meta-filesystems must be included in BBLAYERS (for fuse).
simicsfs must be in DISTRO_FEATURES.
simicsfs must be in MACHINE_FEATURES.
CONFIG_FUSE_FS must be in the kernel config and loaded.

Add a mountpoint for mounting the host at /host.

Change-Id: I6f94bec13fb57432aae894372d4ede1e1d13e6d8
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>