lockout after failed root authentication
This patches Linux-PAM to lock out the root account in the same way as
other accounts: for 5 minutes after 5 failed authentication attempts.

When the account is locked: no authentication attempts to that account
succeed and any attempt will reset the lockout period.   That behavior is
from the pam_tally2 module.

Tested: Yes, testes locking and timeout for both root and non-root.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>