wspoon-tacoma: Patch hwmon for more retries
Increase the amount of I2C read retries that phosphor-hwmon-readd will
do before considering it an error to accommodate mysterious I2C
failures on the max31785 with the AST2600.

On the system I tested on, a few times a minute the max31785 would
require around 10 retries to do a successful read.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: Ia2a89b3d0872c1eed58351f168020bea00228384