meta-ibm: move all meta-witherspoon metadata
The metadata common to all IBM systems is confusingly split between
meta-ibm and meta-witherspoon.  Move everything to meta-ibm.

The meta-ibm README is re-written to more accurately reflect the scope
of the layer.

Allow the configuration samples (bblayers.conf.sample,
local.conf.sample) to be sourced from either meta-ibm or
meta-witherspoon until such a time that any workflows and tooling have
had the opportunity to move to meta-ibm.

Change-Id: Iff9103aebe8cded99dd1e6d1b5a9e3c23b3fbcce
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>