Rainier: Use 55% deviation fan speed deviation
Until the needed fan monitoring algorithm can be implemented that's
necessary to use with the Rainier fan hardware, use a +/-55% deviation
on determining the fan rotors' functional states. The inlet rotor spins
significantly above the requested target whereas the outlet rotor spins
close to the requested target.

Tested: N/A

Change-Id: Ie4cd29485fce6ca1dc7352da718210164daf35b9
Signed-off-by: Matthew Barth <msbarth@us.ibm.com>