meta-ibm: fans: rainier: Add factor & offset to fan monitoring
The inlet rotors of the fan hardware run at a significantly higher speed
above any requested target speed, therefore a factor and offset need to
be applied to these rotors when monitoring their functional states. The
use of a linear equation works for determining the adjusted target that
should be used in determining the functional range of these rotors.
After that is applied, the previous 15% deviation allowed can be used on
all rotors monitored.

The thermal team provided what factor(1.45) and offset(-909) should be
used. These values may change as additional thermal data is collected on
the machine.

Requires: https://gerrit.openbmc-project.xyz/32657

Tested:
    Verified deviation range adjusted for inlet rotors as expected

Change-Id: I04d2368b4b41eea1d4e46ae7393854e9a05cc7e2
Signed-off-by: Matthew Barth <msbarth@us.ibm.com>