meta-ibm: To avoid usb ip is reset issue
We found that the usb0 ip will be cleared when swtiching the ip mode
between DHCP and Static.
As long as the ip starting with "169.254", the ip mode will be regarded
as linklocal.
Because the usb0 ip of mowgli is 169.254.95.120, we need to add this
this special condition to avoid this issue.

Signed-off-by: Ben Pai <Ben_Pai@wistron.com>