host-watchdog: partially re-enable in downstream
The PLDM team is starting to enable host watchdog support so we need to
put this feature back into our downstream code. Until we have full
function in, do nothing if the watchdog expires.

The way the watchdog application is setup to be started by the service,
something must be passed in for the TARGET, so just pass the running
target which will just be a noop.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>