Dump Manager failed to start on whithespoon-tacoma
The hostboot dump directory is created only for p10 systems
because only p10 systems support hostboot dump collection.
But since witherspoon-tacoma supports other OpenPOWER dumps
the hostboot service needs to be started on that too.
The update in this commit is to enable the hostboot dump
directory on whitherspoon-tacoma.
Tests:
Installed the driver with this fix on witherspoon-tacoma
and executed dump tests.

Signed-off-by: Dhruvaraj Subhashchandran <dhruvaraj@in.ibm.com>