mpipl:workaround: do not enable host watchdog
We've enabled the host watchdog in downstream code (sort of) so that the
needed PLDM work can occur in parallel using the watchdog during boot.
The watchdog requires a systemd target to call when it expires so we
just have it currently call the obmc-host-startmin target. This works
fine when it expires during a normal boot but it does not work when the
boot is an MPIPL.

I hate to keep adding hacks but until we can get the official host
watchdog support in, we just need to keep this limping along. For now,
just disable the watchdog in the MPIPL path.

I will remove this along with my other workaround once we have full host
watchdog support in.

This was the partial enable commit:
https://github.ibm.com/openbmc/openbmc/commit/4120c31a02dba8d6d54bdcf745fe1a09da32137e

Fixes SW525729

Tested:
- Verified on a system that the host watchdog did not arm on an MPIPL

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>