meta-phosphor: Add phosphor-boot-indicator recipe
phosphor-boot-indicator indicates BMC boot state via a GPIO. The
package installs a systemd service that looks up and sets the state of
the bmc-management-ready GPIO defined by [1].

The service depends on bmcweb and dropbear. These dependencies are used
as indicators that the system has reached a state that further failures
are able to be inspected either via the web UI, redfish or by inspecting
the BMC from a shell session.

[1] https://gerrit.openbmc-project.xyz/c/openbmc/docs/+/44838

Change-Id: I565303d5b1b2e1bf604a5a3b4e5c4cbace541352
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Eddie James <eajames@linux.ibm.com>