openpower-debug-collector: remove native build support
See openbmc/openpower-debug-collector#2

The move to meson broke something in this path that is not obvious to
me. It's also not clear if this function is really even needed.

For now, disable so we can get back to the SDK building and give us some
time to look further into this.

Tested:
- Verified SDK now builds

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I459ecd7279004c47f89fadc39e224020725c3ae8