Add support for hypervisor network configuration
This commit is created based on the following change in
phosphor-networkd that creates an intermediate app
which takes care of hypervisor network configuration
from bmcweb and and updates the bios manager:
https://gerrit.openbmc-project.xyz/c/openbmc/phosphor-networkd/+/43173

This feature(hyp-nw-config) is IBM specific and when
it is enabled, the service file gets created that starts
after the bios manager service.

It hosts the hypervisor network object that contains eth and
ip address objects which will have corresponding attributes
in the bios table.

The above gerrit commit will give further details on what this
application does.

Tested the same.

Signed-off-by: Asmitha Karunanithi <asmitk01@in.ibm.com>
Change-Id: I5a189b30efe931648a0d71da743ce607107dcfb3