vpnor: Load PNOR TOC lid and ignore missing RO files
The P10 images would not have the PNOR partition files anymore in favor
of lid files, which would be handled by PLDM instead of mboxd. In the
mean time, symlinks are created at runtime for the PNOR partition files,
but they can only be created in the RW directory, therefore ignore
errors if RO files are not found during initialization.

In addition, need to hardcode the lid number for the TOC file, because
this lid is temporary (PLDM doesn't need a TOC file), so it's not listed
in the lids JSON file.

Code that creates the symlinks:
https://gerrit.openbmc-project.xyz/c/openbmc/openpower-pnor-code-mgmt/+/45249

Tested: Verified rainier powered on with image that only had lid files
and the symlinks created.

Change-Id: I28ad0413123215b620acb17cb1048375776c4443
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>