dbus-monitor: add dependency on snmp service
All IBM systems support the SNMP trap notification for error logs. This
means phosphor-dbus-monitor needs to be careful to not start until the
SNMP service is up and running.

Without this dependency, situations can be hit, when lots of errors are
present, where the phosphor-dbus-monitor calls into the SNMP shared
library and hits an unhandled exception when trying to talk to the SNMP
service.

Tested:
- Verified the unhandled exception is no longer seen and
  phosphor-dbus-monitor starts after the SNMP service.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I8b4ca94c0cdd7721aa8847dc478f3827f040a654