Change the service user to the service agent role
This changes the "service" user's role to priv-oemibmserviceagent (was
priv-admin).

This also removes the requirement in the BMC's dropbear SSH server which
only allows connections to accounts which are in the priv-admin group.
That is needed so the service account can connect via SSH.

The service account was added to a new group "shellaccess" so we can re-
enable the SSH access control in a future version.  This control cannot
be re-established now because of forward compatibility needs.  Specifically,
the existing accounts in the readwrite overlay are not in this group, so
after an update they would be denied SSH access.

Tested:
There are two testing modes: new install and upgrade.
This is intended to work in a firmware update scenario, but that
has not been tested.  Testing was performed for a new install (QEMU).

Was able to login via both HTTPS and SSH for both root and service.

The admin account SSH session dropped immediately.
Additional admin accounts dropped their SSH sessions immediately.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>