openpower-debug-collector: Remove watchdog-timeout systemd service link
The openpower-debug-collector-watchdog-timeout@0.service is a link
to openpower-debug-collector-watchdog-timeout@.service. This service
file does not exist in /lib/systemd/system directory. Removing the
entries corresponding to creation and deletion of the link.

Testing:
Installed the image on simics and verified that the systemd service
link corresponding to watchdog-timeout was not setup on the BMC
system.

Signed-off-by: Shantappa Teekappanavar <sbteeks@yahoo.com>
Change-Id: Id6d9309d35bf218a2b458548f8a9dc6b7cc0a58c