Make soft-off service to be a soft dependency for host targets
In the current state soft-off service is a hard dependency
(requires) for the host shut down target, so in any case where
the soft-off service fails with an error condition, the shutdown
target fails too.

This commit intends to make soft-off service as a soft dependency(wants)
so that the shutdown target can still proceed even though the
pldmsoftoff service fails for whatever reason.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>
Change-Id: I5650702c6eb71adab10fdb6493c9d582a9cb0129