witherspoon-tacoma: disable host watchdog function
Hostboot utilizes inband IPMI on witherspoon-tacoma. They relied on OPAL
firmware to continue pinging the watchdog and eventually disabling it.
For the most part though, IBM utilizes PHYP on witherspoon-tacoma
machines and they do not have the needed in-band IPMI support to do
this.

IBM needs either both firmware stacks (hostboot and PHYP) to use PLDM
(i.e. p10bmc) or IPMI (witherspoon). Tacoma is an outlier so just
disable the host watchdog function on Tacoma.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>