ibm: mux-workarounds: Various fixes
Checking for the presence of the mux devices before rebinding defeats
the purpose, since the devices won't probe up the first time without
the reset. Therefore, remove bash exit-on-failure feature, and always
try to re-bind the mux devices.
In addition, the muxes on busses 4, 5, and 6 do not need to be reset,
so remove those.
Finally, make sure the service is done before starting the rest of the
system up, in case there are gpio and i2c dependencies.

Signed-off-by: Eddie James <eajames@linux.ibm.com>
Change-Id: I64426d841914e90f98018d472764be6ecdcd62d4