Enable WebSocket For SOL Hypervisor
Serial Over Lane SOL hypervisor WebSocket is available for the rainier
and Tacoma system.

To build bmcweb with SOL hypervisor WebSocket, we need to enable the
'hypervisor-serial-socket' flag on the bmcweb side.

The default value of 'hypervisor-serial-socket' is disabled because it
only gets introduce for Rainier and Tacoma systems.

This commit enables a build flag for that two systems, using
EXTRA_OEMESON:append:<system>.

Change-Id: Ie6b534383d6d6060392bfeb1c51396106f814b6b