meta-ibm:pstore: remove console from ramoops
The documentation for CONFIG_PSTORE_CONSOLE indicates the following:
"pstore will log all kernel messages, even if no oops or panic happened"

This causes a console-ramoops-0 to appear in /var/lib/systemd/pstore/ on
any BMC reboot. This in turn causes phosphor-debug-collector, which
monitors for any files in this directory, to generate a BMC dump with
the file. This is not what we want.

What we really need are the dmesg-ramoops-* files which are generated
when a kernel panic or oops occur. Remove the CONSOLE setting so we only
get dumps when an actual BMC failure occurs.

Tested:
- Verified that console-ramoops-0 file no longer appears in
  /var/lib/systemd/pstore/ after a BMC reboot.
- Verified we still get needed dmesg-ramoops-* files when a kernel panic
  occurs.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I6b48a8715946b193257bcc4892e07f87c13c95e6