mowgli: disable flash side switch on watchdog (#202)
There is an unknown deadlock occurring intermittently in mowgli when
the host starts. The ideal recovery from this is to simply reset the
BMC and stay on the current side of BMC flash.

Currently, the hardware watchdog triggers a BMC flash switch during
the reboot. This patch removes the flash side switch but keeps the
reboot behavior.

Tested:
- Injected a watchdog hit during host boot and verified BMC came
  back up on same flash side and host was still up and running fine

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>