Add code to upgrade IBM accounts
This changes Linux groups, which builtin users are in these groups, and
how the BMC's dropbear SSH server uses these groups to control access.
This also fixes a bug in the command line to start the dropbear server.

This changes the login shell for admin back to /bin/sh to allow the admin
to SSH port 2200 (host console).

This creates a login profile for the "service" user to drop into a sudo
shell.

This adds new code to upgrade IBM accounts (admin, service, and root)
with the above attributes.

Tested SSH access:
Root user:
 - 22 (BMC command line) allowed
 - 2200 (host console) allowed
 - 2201 (PHYP console) allowed

admin user:
 - 22 (BMC command line) not allowed
 - 2200 (host console) allowed
 - 2201 (PHYP console) not allowed

The disallowed SSH logins get the following messages:
  For the interactive SSH session:
    Permission denied, please try again.
  For the BMC system log:
    dropbear: Logins are restricted to the group shellaccess but
              user 'admin' is not a member

service user:
Authentication is via the ACF upload
 - 22 (BMC command line) allowed, and user running sudo
 - 2200 (host console) allowed
 - 2201 (PHYP console) allowed

Note that initial testing was performed on a system which did not have an
active host console or PHYP console.  So end-to-end testing is needed and
can be performed as a unit test.

Tested service account attribute upgrade script:
This creates the required groups, adds users to groups, etc.
- Script is harmless (makes no changes) on a system which already has
  the required attributes.
- Tested by manually downgrading the system and then running the script.
- Tested by running the script a second time to ensure it is harmless.

Tested the service user has sudo access:
This was tested as part of the service user SSH access.
 - The login prompt is '#' which is for root.
 - The whoami command reports root as the effective user id.
 - Commands are allowed to access resources allowed only to root.
 - Logout works smoothly.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>