Add ibm-ups application to IBM system images
Add the ibm-ups application to IBM system images for Rainier and
Everest.  This application monitors an Uninterruptible Power Supply
(UPS) connected to the system via a USB cable.

This application is being developed downstream because it is a
short-term solution.  Longer term the goal is to use the UPower open
source package to monitor UPS devices.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: Idf048b3b674929e1b270c6df4dd77170c1ccb597