phosphor-fan: Change control/mon dep to .wants
The JSON based fan control and monitor applications previously had a
'requires' dependency for the chassis power on target.  This means that
if they were to die with power on (and have the restarts all fail) that
target would fail and the system would power off.  This occurs even at
steady state host runtime.

This isn't necessary though, so change the dependency to 'wants' instead
of 'requires' so it doesn't happen.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I4fb63ba5d9878c8f85408451b3e458a97e887f09