meta-ibm: add sfclient recipe
sfclient or "signing framework" client is a client application for
interacting with an instance of the IBM verified boot signing framework:

https://github.com/open-power/sb-signing-framework

sfclient also provides a PKCS11 module which is enabled by default,
which enables the use of the framework with any PKCS11 enabled
application.

An SFCLIENT_SIGN_ENABLE variable is provided, intended to be set to "1"
in the shell environment prior to invoking bitbake when performing
builds that wish sign using the signing framework.

sfclient requires libcurl be configured with libssh2 support, which is
not the default.  SFCLIENT_SIGN_ENABLE will enable the correct libcurl
dependency on libssh2 automatically.

Change-Id: Idd25ee818c7863befb4b18b90540c7cfa6dc6cfb
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>