meta-ibm: add sfclient signing for socsec
socsec requires a helper script that reads the hash to be signed from
stdin and prints the signed payload to stdout.  socsec will also want to
verify the signed payload with a key.  Add a helper script
'sfclient-socsec-helper' that adapts sfclient to those requirements.
sfclient-socsec-helper is selected as the helper script by default; it
can be overriden to any arbitrary script by setting
SFCLIENT_SOCSEC_SIGN_HELPER.

sfclient-socsec-helper pulls its configuration from a json configuration
file in the same format as that required by the sfclient pkcs11 module
wrapper.  The path to the configuration is specified in the
SF_PCKS11_CONFIG environment variable prior to invoking bitbake.

sfclient connects remotely to a keyserver and requires a local encrypted
private key to connect, so SSH_AUTH_SOCK and SSH_AGENT_PID are also
required (and the appropriate key previously added to the agent).  Like
SF_PKCS11_CONFIG, these are picked up from the shell environment from
which bitbake is invoked.  Note that while SFCLIENT_SOCSEC_SIGN_HELPER,
SF_PCKS11_CONFIG and SOCSEC_SIGN_KEY can all be changed at any time
after sourcing the OE build environment and will cause a re-sign,
the initial values of SSH_AUTH_SOCK and SSH_AGENT_PID are cached on the
first bitbake invocation in a new build directory, and must be
invalidated manually if the user wishes to switch to a different agent.
Invalidation of the bitbake recipe parse cache can be accomplished by
changing one of the other variables such as SF_PCKS11_CONFIG or manually
deleting the cache directory in the build folder.

Finally, add sfclient as a dependency when deploying u-boot.

Change-Id: Id73e405bdc0f1af9bfbedd13cad58f0b90d5538c
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>