Invoke sudo without explicit shell (#1825)
This fixes a problem where the service user's profile invokes the sudo
command with an explicit invocation of /bin/sh.  Doing so changes the
variables which are present in the ultimate interactive shell's env,
for example the PS1 variable (which controls the command prompt) is not
present.  This corrects the problem by allowing the sudo -i option to
invokke the shell in a way that passes along these variables.

Tested: Yes. With this fix the differences in environment variables
between ssh root@$BMC -vs- ssh service@$BMC are as follows:
- SUDO_GID, SUDO_COMMAND, SUDO_USER, and SUDO_UID are present only for
  service.  This is expected and should be harmless.
- SSH_CONNECTION, SSH_CLIENT, LANG=C, and SSH_TTY are present only in
  the root shell.

Signed-off-by: Joseph Reynolds <joseph-reynolds@charter.net>

Co-authored-by: Joseph Reynolds <joseph-reynolds@charter.net>