Change Model Number to Part Number for FRU products (#1930)
The current implementation in ipmid uses 'Model Number' as a key to
build FRU product info. The FRU YAML file was added in the previous
commit matching the key as well as the D-BUS property, Model, to read
the product info. It appears IBM platforms print 'Part Number'
instead of 'Model Number'. This commit fixes YAML side of the issue.
This fix in FRU YAML and ipmid prints 'Part Number'.

It was tested against our simulation model.
$ ipmitool -I lanplus -C 17 -N 3 -p 2623 -U root -H <host> fru print 1
 Product Name          : Memory DIMM
 Product Part Number   : 78P6575
 Product Serial        : 0001DP911ZZZ

Signed-off-by: Shantappa Teekappanavar <sbteeks@yahoo.com>
Change-Id: Idaf054470a505a8e50cd1d77c9e11029c5a35439

Co-authored-by: Shantappa Teekappanavar <sbteeks@yahoo.com>