meta-ibm: systemd-networkd-wait-online override (#2176)
IBM systems do not always configure both eth0 and eth1 ports. This
causes this service to fail and hold up the boot of the BMC by the 2
minute timeout.

Override this service file to have IBM systems only wait for eth0 OR
eth1 to become active.

Had hoped to just override the ExecStart command but as this is a
oneshot service, ExecStart commands in override files are just appended
so overriding was not an option.

Tested:
- Verified on system with only one network port configured that this now
  quickly exits with success
- Verified that a system with no configured port would still wait on
  this service

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I23c1a6f7417c4f29bf29ab13b6fe122bbd47085e

Co-authored-by: Andrew Geissler <geissonator@yahoo.com>