Enable emergency/rescue targets when root locked (#2257)
When the root account is locked, the systemd emergency and rescue
targets fail with an error message.  That is because they run the
sulogin command, which prompts for the root password.

The solution is for those services to specify the sulogin --force
option.  For more information, see sulogin(8).

systemd uses a "wrapper" executable named systemd-sulogin-shell to run
sulogin.  If the environment variable SYSTEMD_SULOGIN_FORCE is set to 1,
systemd-sulogin-shell runs sulogin with the --force option.  For more
information, see https://systemd.io/ENVIRONMENT/

Create drop-in directories for the emergency and rescue service files.
In those directories, create a .conf file that sets
SYSTEMD_SULOGIN_FORCE to 1 for those services.

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>
Change-Id: I39359e91d99e490b9e57508f96f56567b49c663e