linux-aspeed: New SPI flash controller driver
This is v6 of the patchset to move to a spi-mem based driver. It
replaces the half out of tree driver previously used, and contains
support for the AST2400, AST2500 and AST2600.

 https://lore.kernel.org/linux-mtd/20220503060634.122722-1-clg@kaod.org/

Thank you to Cédric for his hard work maintaining the SPI NOR driver and
reworking it to be a SPI mem driver.

Cédric Le Goater (9):
      ARM: dts: aspeed: Adjust "reg" property of FMC/SPI controllers
      dt-bindings: spi: Add Aspeed SMC controllers device tree binding
      spi: spi-mem: Convert Aspeed SMC driver to spi-mem
      spi: aspeed: Add support for direct mapping
      spi: aspeed: Adjust direct mapping to device size
      spi: aspeed: Workaround AST2500 limitations
      spi: aspeed: Add support for the AST2400 SPI controller
      spi: aspeed: Calibrate read timings
      ARM: dts: aspeed: Enable Dual SPI RX transfers

Potin Lai (1):
      mtd: spi-nor: aspeed: set the decoding size to at least 2MB for AST2600

Tao Ren (1):
      ARM: dts: aspeed-g4: Set spi-max-frequency for all flashes

Change-Id: Ia1bb25a455c0c6f0d8e92ff3241d60bd234a1fa2
Signed-off-by: Joel Stanley <joel@jms.id.au>