ensure vpd collection runs before apr
Recently code was added to move the vpd collection services to later in
the BMC boot. This caused a conflict with the APR (Automatic Power
Restart) service. The APR service would start the power on of the host
before vpd collection had completed. This caused conflicts on the shared
i2c busses between the BMC and host firmware.

The fix here is twofold, try our best to make this APR service run after
vpd collection (this commit) and add a delay to the APR code when it
does need to issue a power on (separate phosphor-state-manager commit).

Tested (with APR delay):
- Powered on p10bmc chassis, caused an AC cycle, and verified that APR
  did not initiate the boot until after vpd collection had completed.
  No host i2c errors were logged and system booted as expected.
- Issued a r/r and verified service did not run
- Verified service exited quickly when no host power on was needed

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>