Update Bmcweb to Add Audit Event Support
This commit will update the bmcweb recipes to allow use of the
audit library. This is change is needed for the addition of
audit logging. These additions should be added to the
downstream openbmc repository because the discussion on
audit logging is still ongoing. There are plans to add
the auditing function upstream, but for now it has been
decided to keep this function downstream.

Link to discussion:
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/55532

Tested:

Compiled with recipe changes. Ensured the necessary libraries
were able to be added.