meta-ibm: vpd: Add wait for VPD Collection (#2767)
The VPD parsers are trigered off by udev events. By nature, this is
asynchronous, the BMC ready state, for example, does not wait until the
VPD collection is complete.

Similarly, on IBM systems, a power down path also triggers udev events
for certain eeproms that are also shared with the host bus. In order to
prevent a poweron when the VPD parsers are still running, we should make
the power off service dependent on the VPD parsers that trigger off as a
result of the udev events to complete.

This commit adds a systemd service that uses the systemctl CLI to
determine if there are any active VPD parser services still running. We
do this every 2 seconds until we see that no VPD parser services are
active, retrying for a maximum of thirty times.

The service is run as a part of the multi-user.target as well as the
power off target.

Tested:
Tested on a few max config systems where the VPD collection would
typically overshoot the BMC multi-user.target. With this service now
added, the multi-user.target is not reached until we see that all VPD
parsers have quiesced.

Change-Id: I561a2c7b65b6139d8431e84f7f83413d31ef6bf8
Signed-off-by: Santosh Puranik <santosh.puranik@in.ibm.com>