meta-phosphor: systemd: remove udev from USERADD_PACKAGES
The latest Yocto adds udev to the USERADD_PACKAGES set, which
causes it to have a RDEPENDS on "shadow".  We explicitly remove
shadow from our initramfs due to not wanting the size of it and
all the corresponding PAM modules.  Having it set breaks p10bmc
where we pull a udev-requiring script into the initramfs.

The only reason udev is currently in the USERADD_PACKAGES is due
to a need for a specific udev rule related to graphics adding a
`render` group.  Since this isn't applicable to the BMC, I'm
removing udev from the USERADD_PACKAGES for us.

Change-Id: I9b92dbefc691f91c179a0bfcb5bd228bb27b738a
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>