Pull in boost 1.81.0
Boost 1.81.0 has some very good changes for us, as it drops the
boost-url beta dependency, in leiu of standard boost.  Because it has
some breaking API changes, it seems like a good idea to get ahead of it,
and pull in the update a few weeks ahead of time.  Looking through the
changelogs, there's nothing in the modules that we use that would pose
any concern for us, aside from the boost-url changes, which is the point
of this commit.

This needs to go hand in hand with:
https://gerrit.openbmc.org/c/openbmc/bmcweb/+/55021

Signed-off-by: Ed Tanous <edtanous@google.com>
Change-Id: I71c629878f386849f8af1e9cb3e6926d3b19f3f9