sysvinit package needed for /run/lock (#3072)
Upstream systemd made the inclusion of the tmpfiles.d/legacy.conf a
feature that is only provided if the sysvinit PACKAGE is set. By default
systemd upstream sets this but OpenBMC overrides that in this file.

The legacy.conf provides among other things, a creation of the
/run/lock directory.

Applications like fw_printenv and fw_setenv rely on the /run/lock
directory being present.

opened systemd/systemd#15668 to see if there could be a more efficient
way of getting /run/lock. Tying it to sysvinit and having it in a file
called legacy.conf just doesn't make a lot of sense to me.

Tested: Verified /run/lock is now present in QEMU boot and that
fw_printenv worked.

Change-Id: I4b4a21eba4bead98365344edebf00a1ee461a57a
Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>