Enable Redfish query expand parameter (#3319)
This enables Redfish expand query parameter. Expand is going to be used
by the GUI on pages like the sensor page to improve performance. There
is some concern about resource usage. I think dropping the MaxLevels to
just 1 or 2 should meet the GUI's usecase and help with resource usage.

https://github.com/openbmc/bmcweb/blob/77665bdac1063ae547890a983daeab7f4320457d/test/redfish-core/lib/service_root_test.cpp#L94
https://github.com/openbmc/bmcweb/blob/6c068982d54d20476fd04ae7ef893b153015762e/redfish-core/lib/service_root.hpp#L95
and a check here:
https://github.com/openbmc/bmcweb/blob/6177a301de5cefdb4a31601ec2d899f4309fc6c2/redfish-core/include/utils/query_param.hpp#L280

Signed-off-by: Gunnar Mills <gmills@us.ibm.com>