Enable Audit Logging using auditd
bmcweb issues audit events to be tracked when audit-events is enabled.
(Note: the build depends on the audit recipe, while the runtime depends
on the auditd package.)

Configure auditd to create 2MB log files and keep at most 5 of them. No
special handling is performed for running out of file system space or
errors in writing the log file.

Turning on auditd also turns on auditing for the kernel. The BPF
subsystem floods the log with many entries. Rules are added to exclude
these events from being logged.

The audit events are also by default also going to
systemd-journald-audit.socket and reported by journalctl. We do not want
duplication of these events. Turned off systemd-journald-audit.socket to
prevent the duplication.

Tested: Installed on both Simics and real hardware. Wrote a script to
cause multiple log files to be created and oldest ones to be deleted.
Ran various Redfish API's through bmcweb and confirmed event was logged.