Adjust dump size to account for audit log additions
The audit log will be included in the bmc dump. The default
configuration allows up to 10MB of audit log files to be created.
Adjusting the maximum size of the bmc dump to account for this
additional data from 20MB to 25MB. (This is compressed, so don't need
additional full 10MB.)

Signed-off-by: Janet Adkins <janeta@us.ibm.com>