meta-ibm:ncsi: ensure eth speeds set to 1 gigabit (#3457)
IBM has run into some issues recently with certain network
configurations where eth1 is not getting set to 1 gigabit. This can
cause issues in scenarios where a network environment is running on 1
gigabit minimum-speed network switches.

Removing link-local support from eth1 seemed to aggravated this.
link-local is only on eth0 to avoid subnet issues. But without
link-local, the systemd call, systemd-networkd-wait-online, does not
always find an active eth1 so it never runs the command to force eth1 up
to 1 gigabit.

As far as we can tell, there's no appropriate mapping within
systemd-networkd-wait-online for it to know when the NCSI stack is
appropriately up on a network interface (allowing the 1 gigabit speed to
be set). Therefore the solution taken in this commit is to just keep
retrying for a period of time.

Tested:
- Verified new script ran and set eth ports as expected

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I6737e208b59830262f51a8e30c78a97fc698e4fd

Co-authored-by: Andrew Geissler <geissonator@yahoo.com>