Revert "Add socket-id for the first console"
This reverts commit ec7cab9378f548e94ee4af83946b2583d1fc7194.

ec7cab9378f5 ("Add socket-id for the first console") was only a partial
fixup of the obmc-console configurations found in the tree. For
historical reasons not all platforms that support multiple consoles
install a client configuration file for all consoles. Instead they relied
on some default behaviour that was removed in obmc-console's 4e7186918599
("Fixed broken dbus interface for multiple consoles"). Applying the
configuration changes in the manner of ec7cab9378f5 ("Add socket-id for
the first console") resulted in an asymmetric configuration between the
client and server, breaking some SSH SOL instances.

As of ae2460d0b8e8 ("obmc-console: Provide a default value for
`console-id`.") in obmc-console the requirement to specify
`socket-id`[^1] is lifted. Instead, the configuration can choose to
override a default value. This restores the original behaviour and
unbreaks SSH SOL.

[^1]: Now an alias of the `console-id` configuration key

Websocket and IPMI consoles remain broken. A fix is under development
that also takes a proper approach to supporting multiple consoles in
bmcweb, and the same solution pattern can be applied in ipmid.

More discussion of the problems involved can be found here:

https://amboar.github.io/notes/2023/05/08/happenings-in-obmc-console.html

Change-Id: I274284c791758f336da6d6301dc523bac2b5dd69
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>