meta-phosphor: pam-ipmi: prep for usrmerge
Change If53ea402f1e50416bc8a0e561d42080ea68742f9 is going to require
us to pass the `--libdir` configure option so that the install
directory of the PAM modules is not hard-coded, but instead relies on
the `${base_libdir}` derived by the usrmerge feature.  Add this now in
preparation.

This is the same approach taken by the base PAM recipe.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I1a46177f65347521debca54f00145d32872b05b6