1050-ghe: revert power restore delay (#3668)
To keep forward compatibility, we can not allow new d-bus properties to
be added to existing d-bus interfaces. The following issue describes the
problem:
  https://github.com/openbmc/phosphor-settingsd/issues/16

In summary, when a new property is added to an existing interface, any
old properties do not have their value preserved upon a firmware update
to a level with the new property. This causes serious issues within PSM
as the one-time APR policy is utilized for inband code updates. This
issue causes inband code updates to fail.

We will have to go back to the 1020-1040 solution which was to just hard
code a 30s delay.

This commit includes the needed bumps to the ibm-openbmc PSM and PDI
repos to pick up the reverts and old 1040 implementations.

Tested:
- Flashed a 1040 driver, set the one-time APR to AlwaysOn, updated to an
image with these changes and confirmed that after the update the auto
power on logic started a boot.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>

Co-authored-by: Andrew Geissler <geissonator@yahoo.com>