meta-ibm: p10bmc: Move Bonnell dtb to be first on list (#3773)
The first dtb listed in the KERNEL_DEVICETREE list is the default used
for the BMC's genesis boot. The Bonnell device tree specifies an eeprom
that's larger in size than other p10bmc systems. Therefore the device
driver currently truncates the data on this eeprom when doing a genesis
boot on Bonnell since the default device tree is for Rainier.

Move Bonnell to the top of the list to be the default since it has the
larger eeprom size.

Tested: Booted Rainier and Everest on simulation to test a genesis BMC
boot. Verified they first booted with the Bonnell's device tree, then
the BMC automatically rebooted after the machine type was identified and
the device tree corresponding to the system was loaded.

Change-Id: Icb9abe7d8b04d42befd9eb906d095f4fa046ab3c
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>