Enable IPv6 AutoConfig/SLAAC by default (#3803)
* Enable IPv6 AutoConfig/SLAAC by default

Autoconfig/SLACC is set to false on both interfaces by default for
now. Updating it to false as expected.

Tested by:

Validated the default value for default-ipv6-accept-ra parameter is
set to true in the network configuration for both interfaces.

Change-Id: I048918726c656b7e02f8dbc43a401f4ae38fbbeb
Signed-off-by: Jishnu CM <jishnunambiarcm@duck.com>

* Enable IPv6 AutoConfig/SLAAC by default

Autoconfig/SLACC is set to false on both interfaces by default for
now. Updating it to false as expected.

Tested by:

Validated the default value for default-ipv6-accept-ra parameter is
set to true in the network configuration for both interfaces.

Change-Id: I048918726c656b7e02f8dbc43a401f4ae38fbbeb
Signed-off-by: Jishnu CM <jishnunambiarcm@duck.com>

Co-authored-by: Jishnu CM <jishnunambiarcm@duck.com>