meta-aspeed: Fix value_start issue in `rev_id` path
There's a bug in socsec that prevents the IBM's configuration from
successfully building an OTP image.

I have got the fix merged upstream:

https://github.com/AspeedTech-BMC/socsec/pull/18

However, Aspeed do not plan on doing a release until October:

> [ 17:51 ] arj: @Troy Lee any chance Neal can tag a new socsec release so we can bump it in OpenBMC and pick up some recent fixes?
> [ 18:25 ] Troy Lee: Current schedule is October.

https://discord.com/channels/775381525260664832/922871693008068638/1144547174286377062

For now, fix otptool using a recipe patch, in violation of the usual
guidelines.

I prefer we do this over switching to a "git" version for the recipe as
instability with these tools really cannot be tolerated.

Change-Id: I65b1992b5479ea257cfa65fd8b3cfc021b7d3dea
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>