meta-ibm: simplify phosphor-health-monitor config
The latest phosphor-health-monitor code has a default config and merges
it with the provided config.  Therefore you do not need to express
configs that do not need changes.

The only change the IBM systems have is that they monitor `/var` instead
of `/run/initramfs/rw`, due to the flash layout differences.  Provide
the minimal config for that.

Tested:

Modify phosphor-health-monitor to point at this config file and ran it
on my desktop.  Observed the expected dbus sensor tree.
```
$ busctl --user tree xyz.openbmc_project.HealthMon
└─ /xyz
  └─ /xyz/openbmc_project
    └─ /xyz/openbmc_project/metric
      └─ /xyz/openbmc_project/metric/bmc
        ├─ /xyz/openbmc_project/metric/bmc/cpu
        │ ├─ /xyz/openbmc_project/metric/bmc/cpu/kernel
        │ ├─ /xyz/openbmc_project/metric/bmc/cpu/total
        │ └─ /xyz/openbmc_project/metric/bmc/cpu/user
        ├─ /xyz/openbmc_project/metric/bmc/memory
        │ ├─ /xyz/openbmc_project/metric/bmc/memory/available
        │ ├─ /xyz/openbmc_project/metric/bmc/memory/buffered_and_cached
        │ ├─ /xyz/openbmc_project/metric/bmc/memory/free
        │ ├─ /xyz/openbmc_project/metric/bmc/memory/shared
        │ └─ /xyz/openbmc_project/metric/bmc/memory/total
        └─ /xyz/openbmc_project/metric/bmc/storage
          ├─ /xyz/openbmc_project/metric/bmc/storage/rw
          └─ /xyz/openbmc_project/metric/bmc/storage/tmp
```

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
Change-Id: I0ff8528fbb8d9cbf2ad4c6672d5a1dc11cf5a970