phosphor-power: Complete removal of ibm-ups app (#5169)
Due to inter-repository dependencies, a temporary meson setting needed
to be added to the phosphor-power recipe.  This was required until the
ibm-ups application was removed from the phosphor-power repository.

The application has now been removed from phosphor-power, so remove the
temporary meson setting from the recipe.  The setting is no longer
necessary or valid.

Tested:
* Built phosphor-power in bitbake
* Built p10bmc image in bitbake

Change-Id: I4045ac57098e3fa963d5380d68e8023102c98aab

Signed-off-by: Shawn McCarney <shawnmm@us.ibm.com>