meta-ibm:system1: disable reading of non-i2c vpd chips
System1 has a lot of i2c mux's and devices behind those mux's. The
entity manager fru device ends up scanning devices which are shared
with the host firmware depending on the mux settings. The only thing fru
device needs to be looking at are the vpd chips so exclude everything
else on the bus's.

Tested:
- Confirmed vpd is still read and no bus busy errors are logged anymore

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>