smbios-mdr:Add support to toggle pcieslot presence (#5769)
In the current state, the PCIeslot presence is always hardcoded to
`true`, since the PCIeSlots are always embedded on the board. But the
slot presence is not really useful. Since we are already getting the
current usage of the slot in the SMBIOS Type 9 structure we could use
the "Available"/"Occupied" status in the current usage field to
indicate the drive presence.

Modify the slot presence code to only mark "true" if the current usage
of the slot is "Occupied" & make it "false" if it is anything other than
that.

Signed-off-by: Manojkiran Eda <manojkiran.eda@gmail.com>
Co-authored-by: Manojkiran Eda <manojkiran.eda@gmail.com>